--- multegulaUI/easy_high_scores.txt
@@ -1,10 +1,0 @@
-Nov. 18, Eats - 4998
-Nov. 29, Primanti - 4997
-May. 09, ToRo - 4297
-No Date Found, Bonnerjam - 3928
-Feb. 12, Unknown - 1897
-No Date Found, DThomp - 207
-Mar. 01, hello - 176
-Nov. 30, BDubs - 140
-Nov. 27, TexasRH - 102
-Nov. 30, TheO - 49
