--- multegulaUI/hard_high_scores.txt
@@ -1,10 +1,0 @@
-Nov. 18, Friends - 6897
-Dec. 01, Beesi - 6896
-Nov. 29, David - 4337
-Nov. 27, Chris - 4119
-Nov. 23, KP - 3139
-Nov. 28, Alex - 2684
-Nov. 24, Spencer - 2588
-Nov. 22, Kevin - 2549
-Nov. 29, Jenn - 2476
-Nov. 22, Elise - 2055
