--- multegulaUI/normal_high_scores.txt
@@ -1,10 +1,0 @@
-Nov. 18, Bands - 5467
-Nov. 30, AVA - 5166
-Nov. 27, Coldplay - 4146
-Nov. 27, Future - 4139
-Nov. 30, of - 2175
-Nov. 23, Forestry - 2139
-Nov. 24, U2 - 1588
-Nov. 22, Dashboard - 1549
-Nov. 27, Emery - 1441
-Nov. 27, Jack - 1341
