+++ multegulaUI/screens/Screens.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class Screens(Enum):
+  SCRN_NONE = 0;
+  SCRN_MAIN = 1;
+  SCRN_PAUSE = 2;
+  SCRN_GAME = 3;
